Refactor wallpaper setting feature, extracting to it's own class.
